Small main class/method
